refactor: adjust the file struct
